splut technicalOptionTab.tsx in multiple files
